% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{column_formater}
\alias{column_formater}
\title{Title}
\usage{
column_formater(fixed_digits = 2, format_str_numbers = TRUE,
  sub_super_convert = TRUE, line_br_to_cell = TRUE,
  latex_converts = TRUE, na = "")
}
\arguments{
\item{na}{}
}
\description{
Title
}
