% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure.R
\name{latex_figure}
\alias{latex_figure}
\title{Title}
\usage{
latex_figure(files, pics_per_row, max_rows_per_fig, fig_caption = NA,
  fig_label = NA, file_captions = NULL, file_labels = NULL,
  subcaption_pos = c("top", "bottom"), fig_scale = 1,
  fig_fixed_width = NA, clearpage = TRUE, centering = TRUE,
  continued_text = "(continued)", add_env_code = NA,
  last_figs_left_aligned = TRUE, smart_layout = TRUE)
}
\arguments{
\item{last_figs_left_aligned}{}
}
\description{
Title
}
