% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{latex_labelmaker}
\alias{latex_labelmaker}
\title{Title}
\usage{
latex_labelmaker(prefix, type = c("fig", "subfig", "tab"), sep = "_",
  counter = c("num", "letter"))
}
\arguments{
\item{counter}{}
}
\description{
Title
}
