% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{latex_table}
\alias{latex_table}
\title{Title}
\usage{
latex_table(df, caption = NA,
  caption_cont_label = latex_it("(continued)"), label = NA,
  alignment = NULL, landscape = FALSE, font_size = NA,
  bf_header = TRUE,
  header_formater = column_formater(format_str_numbers = FALSE),
  top_row = NULL, footer = NULL, col_formater = column_formater(),
  write = c("cat", "str", "clipr"))
}
\arguments{
\item{write}{}
}
\description{
Title
}
