% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{make_table_footer}
\alias{make_table_footer}
\title{Title}
\usage{
make_table_footer(general = NULL, alpha = NULL, number = NULL,
  order = c("general", "number", "alpha"), general_title = NA,
  number_title = NA, alpha_title = NA, font_size = NULL)
}
\arguments{
\item{font_size}{}
}
\description{
Title
}
